requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==3.2
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jsonfield==3.1.0
outcome==1.2.0
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
whitenoise==6.2.0
wsproto==1.2.0
